Refactor: added negative return number to tests that fail

+++ include/fused_kernel/utils/type_lists.cuh
@@ -141,7 +141,7 @@ namespace fk { // namespace fused kernel
         T at[SIZE];
     };
 
-    template <typename T, int BATCH, typename... Types>
+    template <typename T, size_t BATCH, typename... Types>
     FK_HOST_DEVICE_CNST Array<T, BATCH> make_array(Types... pars) {
         static_assert(sizeof...(Types) == BATCH, "Too many or too few elements for the array size.");
         static_assert(std::disjunction_v<std::is_same<T, Types>...>, "All the types should be the same");
